+++ src/components/splitButton/index.js
@@ -1,0 +1,4 @@
+import SplitButton from './SplitButton';
+
+export default SplitButton;
+export { SplitButton };